2019/11 (#53)
* 2019/11 complete

* someday i will comply with my own lint standards

+++ 2019/lib/src/io.c
@@ -24,7 +24,7 @@ int read_csv_line(const char *filename, long *buffer, int bufsize)
     while (tok != NULL)
     {
         assert(count < bufsize);
-        buffer[count++] = atoi(tok);
+        buffer[count++] = atol(tok);
         tok = strtok(NULL, ", \n");
     }
